another text rendering fixes * FontAtlas:   * eternal fix for "ascender" and "offset_y" using FT_Outline_Get_CBox() function.   * do not fallback to 64px font size if fails.   * use "FT_LOAD_IGNORE_TRANSFORM" to speed-up glyph rendering. * FontType:   * rewrite fonttype_measure_line_height() function.   * rewrite fonttype_map_codepoints() function, avoid duplicate codepoints fetch.   * scale paragraph_space before apply.

+++ src-dreamcast/include/externals/fontatlas.h
@@ -42,7 +42,7 @@ typedef struct _FONT_STRUCTS_ATTRIB {
     uint16_t texture_width;
     uint16_t texture_height;
     int32_t texture_byte_size;
-    int16_t ascender;
+    float ascender;
     int16_t line_height;
     FontKerning* kernings_array;
 } FontCharMap;
